% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF.R
\name{convGDX2MIF}
\alias{convGDX2MIF}
\title{Read in GDX and write *.mif reporting}
\usage{
convGDX2MIF(
  gdx,
  gdx_ref = NULL,
  file = NULL,
  scenario = "default",
  time = as.numeric(readGDX(gdx, name = "t"))
)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{gdx_ref}{reference-gdx for policy costs, a GDX as created by readGDX, or the file name of a gdx}

\item{file}{name of the mif file which will be written, if no name is
provided a magpie object containing all the reporting information is
returned}

\item{scenario}{scenario name that is used in the *.mif reporting}

\item{time}{temporal resolution of the reporting, default: c(seq(2010,2050,5))}
}
\description{
Read in all information from GDX file and create
the *.mif reporting
}
\examples{

\dontrun{convGDX2MIF(gdx,gdx_ref,file="LIMES_generic_default.csv",scenario="default")}

}
\author{
Sebastian Osorio and Renato Rodrigues
}
