% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convGDX2MIF_tau.R
\name{convGDX2MIF_tau}
\alias{convGDX2MIF_tau}
\title{Read in GDX and write *.mif time slice reporting}
\usage{
convGDX2MIF_tau(
  gdx,
  file = NULL,
  scenario = "default",
  time = as.numeric(readGDX(gdx, name = "t"))
)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{file}{name of the mif file which will be written, if no name is
provided a magpie object containing all the reporting information is
returned}

\item{scenario}{scenario name that is used in the *.mif reporting}

\item{time}{temporal resolution of the reporting, default: c(seq(2010,2050,5))}
}
\description{
Read in all information from GDX file and create
the *.mif reporting for the power generation at the time slice resolution
}
\examples{
\dontrun{
convGDX2MIF_tau(gdx, file = "LIMES_tau_default.csv", scenario = "default")
}

}
\author{
Antoine Levesque
}
