% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importParameters.R
\name{importParameters}
\alias{importParameters}
\title{Read in GDX and import different parameters used in convGDX2MIF.R for intermediate calculations in
the reporting}
\usage{
importParameters(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in GDX and import different parameters used in convGDX2MIF.R for intermediate calculations in
the reporting
}
\examples{

\dontrun{importParameters(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
