% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limes-package.R
\docType{package}
\name{limes-package}
\alias{limes}
\alias{limes-package}
\alias{remind}
\title{The LIMES R package}
\description{
Contains the LIMES-specific routines for data and model output manipulation
}
\details{
\tabular{ll}{ Package: \tab remind\cr Type: \tab Package\cr Version: \tab
7.6\cr Date: \tab 2017-11-20\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\author{
Sebastian Osorio and Renato Rodrigues

Maintainer: Sebastian Osorio <sebastian.osorio@pik-potsdam.de>
}
