% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesAllocateYears.R
\name{limesAllocateYears}
\alias{limesAllocateYears}
\title{Allocating years to variables that do not change over time}
\usage{
limesAllocateYears(var, gdx)
}
\arguments{
\item{var}{variable to be changed}

\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}
}
\description{
Allocating years to variables that do not change over time
}
\examples{

\dontrun{limesAllocateYears(var)}

}
\author{
Sebastian Osorio
}
