% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesDiscount.R
\name{limesDiscount}
\alias{limesDiscount}
\title{Discounting variables (matrix)}
\usage{
limesDiscount(var, gdx)
}
\arguments{
\item{var}{variable to be changed}

\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}
}
\description{
Discounting variables (matrix)
}
\examples{

\dontrun{limesDiscount(var)}

}
\author{
Sebastian Osorio
}
