% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesInt2Ext.R
\name{limesInt2Ext}
\alias{limesInt2Ext}
\title{mapping of weights for the variables for global aggregation}
\usage{
limesInt2Ext(gdx, output, reporting_tau = FALSE, mappingPath = NULL)
}
\arguments{
\item{gdx}{a GDX as created by readGDX, or the file name of a gdx}

\item{output}{variable to be changed}

\item{reporting_tau}{boolean determining whether to generate the tau reports}

\item{mappingPath}{path to mapping file}
}
\description{
mapping of weights for the variables for global aggregation
}
\author{
Sebastian Osorio and Renato Rodrigues
}
