% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limesMapping.R
\name{limesMapping}
\alias{limesMapping}
\title{Mapping limes iso 2 codes to iso 3 codes}
\usage{
limesMapping(var, mappingPath = NULL)
}
\arguments{
\item{var}{variable to be changed}

\item{mappingPath}{path to mapping file}
}
\description{
Mapping limes iso 2 codes to iso 3 codes
}
\examples{

\dontrun{LIMESMapping(var,mappingPath="R/mapping/mapping.csv")}

}
\author{
Sebastian Osorio and Renato Rodrigues
}
