% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelstat.R
\name{modelstat}
\alias{modelstat}
\title{modelstat}
\usage{
modelstat(gdx, file = NULL)
}
\arguments{
\item{gdx}{GDX file}

\item{file}{a file name the output should be written to using write.magpie}
}
\value{
A MAgPIE object containing the modelstat
}
\description{
LIMES model status of optimization
}
\examples{

  \dontrun{
    x <- modelstat(gdx)
  }

}
\author{
David Klein
}
