% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAdequacyContribution.R
\name{reportAdequacyContribution}
\alias{reportAdequacyContribution}
\title{Read in GDX and calculate capacities, used in convGDX2MIF.R for the reporting}
\usage{
reportAdequacyContribution(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the Capacity Adequacy variables
}
\description{
Read in Capacity Adequacy information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportCapacity Adequacy(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
