% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportAnnualAvFactor.R
\name{reportAnnualAvFactor}
\alias{reportAnnualAvFactor}
\title{Read in GDX and report the annual availability factor (for nonRES), used in convGDX2MIF.R for the reporting}
\usage{
reportAnnualAvFactor(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the annual availability factors (for nonRES)
}
\description{
Read in annual availability factors (for nonRES) information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportAnnualAvFactor(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
