% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBuildings.R
\name{reportBuildings}
\alias{reportBuildings}
\title{Read in GDX and calculate buildings variables, used in convGDX2MIF.R for the reporting}
\usage{
reportBuildings(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains buildings module variables
}
\description{
Read in (net and gross) demand data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportBuildings(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
