% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportBuildtime.R
\name{reportBuildtime}
\alias{reportBuildtime}
\title{Read in GDX and report the plant buildtime, used in convGDX2MIF.R for the reporting}
\usage{
reportBuildtime(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the buildtimes
}
\description{
Read in buildtime information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportBuildtime(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
