% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCapitalCosts.R
\name{reportCapitalCosts}
\alias{reportCapitalCosts}
\title{Read in GDX and report fuel costs, used in convGDX2MIF.R for the reporting}
\usage{
reportCapitalCosts(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the fuel costs
}
\description{
Read in fuel costs information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportFuelCosts(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
