% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCarbonSequestration.R
\name{reportCarbonSequestration}
\alias{reportCarbonSequestration}
\title{Read in GDX and calculate carbon sequestred,  used in convGDX2MIF.R for the reporting}
\usage{
reportCarbonSequestration(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX,  or the path to a gdx}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in carbon sequestred data from GDX file,  information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportCarbonSequestration(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio,  Renato Rodrigues
}
