% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportCurtailment.R
\name{reportCurtailment}
\alias{reportCurtailment}
\title{Read in GDX and calculate electricity curtailment, used in convGDX2MIF.R for the reporting}
\usage{
reportCurtailment(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the curtailment
}
\description{
Read in electricity curtailment data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportCurtailment(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
