% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportElectricalEfficiency.R
\name{reportElectricalEfficiency}
\alias{reportElectricalEfficiency}
\title{Read in GDX and report the electrical efficiency, used in convGDX2MIF.R for the reporting}
\usage{
reportElectricalEfficiency(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the electrical efficiency
}
\description{
Read in electrical efficiency information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportElectricalEfficiency(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
