% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportElectricityPrices.R
\name{reportElectricityPrices}
\alias{reportElectricityPrices}
\title{Read in GDX and calculate electricity prices,  used in convGDX2MIF.R for the reporting}
\usage{
reportElectricityPrices(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX,  or the path to a gdx}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in electricity prices information from GDX file,  information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportElectricityPrices(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio,  Renato Rodrigues
}
