% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportExchange.R
\name{reportExchange}
\alias{reportExchange}
\title{Read in GDX and calculate electricity exchange,  value of exchange and aggregated transmission capacites,  used in convGDX2MIF.R for the reporting}
\usage{
reportExchange(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX,  or the path to a gdx}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in electricity exchange data from GDX file,  information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportExchange(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio,  Renato Rodrigues
}
