% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportFictitiousVars.R
\name{reportFictitiousVars}
\alias{reportFictitiousVars}
\title{Read in GDX and calculate emissions, used in convGDX2MIF.R for the reporting}
\usage{
reportFictitiousVars(gdx, output = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}
}
\value{
MAgPIE object - contains the emission variables
}
\description{
Create fictitious variables to ensure symetry of the magpie object, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportEmissions(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
