% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGeneration.R
\name{reportGeneration}
\alias{reportGeneration}
\title{Read in GDX and calculate electricity generation, used in convGDX2MIF.R for the reporting}
\usage{
reportGeneration(gdx, output = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}
}
\value{
MAgPIE object - contains the Generation variables
}
\description{
Read in electricity generation data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportGeneration(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
