% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportHydrogen.R
\name{reportHydrogen}
\alias{reportHydrogen}
\title{Read in GDX and calculate different variables related to hydrogen production and trade
, including also prices, used in convGDX2MIF.R for the reporting}
\usage{
reportHydrogen(gdx, output = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions
reporting at the time slice level (TRUE) or at the yearly level (FALSE, default)}
}
\value{
MAgPIE object - contains the Generation variables
}
\description{
Read in hydrogen-related data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{
\dontrun{
reportHydrogen(gdx)
}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
