% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportIndustryEmissions.R
\name{reportIndustryEmissions}
\alias{reportIndustryEmissions}
\title{Read in GDX and calculate industry emissions, used in convGDX2MIF.R for the reporting}
\usage{
reportIndustryEmissions(gdx, output = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}
}
\value{
MAgPIE object - contains the emission variables
}
\description{
Read in emissions data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportEmissions(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
