% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInput.R
\name{reportInput}
\alias{reportInput}
\title{Read in GDX and calculate availability factors,  used in convGDX2MIF.R for the reporting}
\usage{
reportInput(gdx, mappingPath = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX,  or the path to a gdx}

\item{mappingPath}{path to mapping file}
}
\value{
MAgPIE object - contains the availability factors
}
\description{
Read in availability factors information from GDX file,  information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportNuren(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio,  Renato Rodrigues
}
