% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportInvestmentCosts.R
\name{reportInvestmentCosts}
\alias{reportInvestmentCosts}
\title{Read in GDX and report the investment costs, used in convGDX2MIF.R for the reporting}
\usage{
reportInvestmentCosts(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the investment costs
}
\description{
Read in investment costs information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportInvestmentCosts(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
