% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportLifetime.R
\name{reportLifetime}
\alias{reportLifetime}
\title{Read in GDX and report the plant lifetime, used in convGDX2MIF.R for the reporting}
\usage{
reportLifetime(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the lifetimes
}
\description{
Read in lifetime information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportLifetime(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
