% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportNuren.R
\name{reportNuren}
\alias{reportNuren}
\title{Read in GDX and calculate availability factors, used in convGDX2MIF.R for the reporting}
\usage{
reportNuren(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the availability factors
}
\description{
Read in availability factors information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportNuren(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
