% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportOMF.R
\name{reportOMF}
\alias{reportOMF}
\title{Read in GDX and report the Fixed O&M costs, used in convGDX2MIF.R for the reporting}
\usage{
reportOMF(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the Fixed O&M costs
}
\description{
Read in Fixed O&M costs information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportOMF(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
