% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportOMV.R
\name{reportOMV}
\alias{reportOMV}
\title{Read in GDX and report the Variable O&M costs, used in convGDX2MIF.R for the reporting}
\usage{
reportOMV(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the Variable O&M
}
\description{
Read in Variable O&M information from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportOMV(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
