% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportOtherEUETSsector.R
\name{reportOtherEUETSsector}
\alias{reportOtherEUETSsector}
\title{Read in GDX and calculate industry emissions, used in convGDX2MIF.R for the reporting}
\usage{
reportOtherEUETSsector(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the emission variables
}
\description{
Read in emissions data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportOtherEUETSsector(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
