% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPeakDemand.R
\name{reportPeakDemand}
\alias{reportPeakDemand}
\title{Read in GDX and calculate annual peak demand, used in convGDX2MIF.R for the reporting}
\usage{
reportPeakDemand(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in gross demand data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportPeakDemand(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
