% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportPrimaryEnergy.R
\name{reportPrimaryEnergy}
\alias{reportPrimaryEnergy}
\title{Read in GDX and calculate primary energy, used in convGDX2MIF.R for the reporting}
\usage{
reportPrimaryEnergy(gdx)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}
}
\value{
MAgPIE object - contains the capacity variables
}
\description{
Read in primary energy data from GDX file, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportPrimaryEnergy(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio, Renato Rodrigues
}
