% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportUKETSvars.R
\name{reportUKETSvars}
\alias{reportUKETSvars}
\title{Read in GDX and calculate emissions, used in convGDX2MIF.R for the reporting}
\usage{
reportUKETSvars(gdx, output = NULL)
}
\arguments{
\item{gdx}{a GDX object as created by readGDX, or the path to a gdx}

\item{output}{a magpie object containing all needed variables generated by other report*.R functions}
}
\value{
MAgPIE object - contains the emission variables
}
\description{
Read in data that only exists at EU ETS level, information used in convGDX2MIF.R
for the reporting
}
\examples{

\dontrun{reportUKETSvars(gdx)}

}
\seealso{
\code{\link{convGDX2MIF}}
}
\author{
Sebastian Osorio
}
